requirements.txt
anthropic==0.28.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
sqlite3-python==1.0.0
pydantic==2.5.0
tqdm==4.66.1
